more piston scripts and moving object
Made the other pistons able to move, as well as a square that moves along the conveyor belt. The green and blue pistons can change the colour of the square too. The red piston doesn't do anything yet, neither does the button i added on the left of the screen.